Fix read of incorrect value while parceling

+++ plugins/openpgp-api-lib/openpgp-api/src/main/java/org/openintents/openpgp/util/OpenPgpKeyPreference.java
@@ -263,7 +263,7 @@ public class OpenPgpKeyPreference extends Preference implements OpenPgpApiManage
         public SavedState(Parcel source) {
             super(source);
 
-            keyId = source.readInt();
+            keyId = source.readLong();
             defaultUserId = source.readString();
         }
